Use StreamProcessor protocol, not ess.reduce StreamProcessor impl
